+++ src/compositions/bcl-group/data/data.js
@@ -111,51 +111,6 @@ module.exports = {
       },
     ],
   },
-  pagination: {
-    attributes: new drupalAttribute().addClass(["mt-4"]),
-    alignment: "center",
-    border_top: true,
-    prev: {
-      label: "Previous",
-      path: "#",
-    },
-    next: {
-      label: "Next",
-      path: "#",
-    },
-    items: [
-      {
-        label: "1",
-        path: "#",
-        active: true,
-      },
-      {
-        label: "2",
-        path: "#",
-      },
-      {
-        label: "3",
-        path: "#",
-      },
-    ],
-  },
-  filter_button: {
-    label: "Filter options",
-    type: "button",
-    variant: "light",
-    size: "lg",
-    attributes: new drupalAttribute()
-      .addClass(["d-lg-none", "w-100", "mb-3-5"])
-      .setAttribute("data-bs-toggle", "offcanvas")
-      .setAttribute("data-bs-target", "#bcl-offcanvas")
-      .setAttribute("aria-controls", "bcl-offcanvas"),
-    icon_position: "before",
-    icon: {
-      name: "filter",
-      path: "/icons.svg",
-      size: "xs",
-    },
-  },
   badges: [
     {
       label: "Group name",
@@ -177,17 +132,4 @@ module.exports = {
       attributes: new drupalAttribute().addClass(["me-2", "text-dark", "mb-2"]),
     },
   ],
-  sort_select: {
-    label: "Sort&nbsp;by",
-    attributes: new drupalAttribute().addClass("mb-4 mb-md-0 ms-md-4"),
-    aria_label: "sort-select",
-    label_extra_classes: "mb-0",
-    options: [
-      { value: 1, label: "a select option" },
-      { value: 2, label: "another select option" },
-      { value: 3, label: "another option" },
-      { value: 4, label: "last option" },
-      { label: "Joined date", selected: true },
-    ],
-  },
 };
