+++ src/compositions/bcl-group/group.story.js
@@ -7,15 +7,18 @@ import isChromatic from "chromatic/isChromatic";
 import { withDesign } from "storybook-addon-designs";
 import header from "@openeuropa/bcl-data-header/data--simple";
 import footer from "@openeuropa/bcl-data-footer/data";
+import drupalAttribute from "drupal-attribute";
+import dataListingTemplate from "@openeuropa/bcl-data-templates/listing/data--template";
+import dataProfiles from "@openeuropa/bcl-data-templates/listing/data--profiles";
+import listingPage from "@openeuropa/bcl-base-templates/listing-page.html.twig";
+
 import groupListing from "@openeuropa/bcl-group/group.html.twig";
 import groupLanding from "@openeuropa/bcl-group/group-landing.html.twig";
-import listingPage from "@openeuropa/bcl-base-templates/listing-page.html.twig";
-import demoData from "@openeuropa/bcl-group/data";
-import demoDataList from "@openeuropa/bcl-group/dataList";
-import demoDataListContent from "@openeuropa/bcl-group/dataListContent";
-import demoDataListMember from "@openeuropa/bcl-group/dataListMember";
-import demoDataLanding from "@openeuropa/bcl-group/dataLanding";
-import drupalAttribute from "drupal-attribute";
+import demoData from "@openeuropa/bcl-group/data/data";
+import demoDataList from "@openeuropa/bcl-group/data/dataList";
+import demoDataListContent from "@openeuropa/bcl-group/data/dataListContent";
+import demoDataListMember from "@openeuropa/bcl-group/data/dataListMember";
+import demoDataLanding from "@openeuropa/bcl-group/data/dataLanding";
 
 if (isChromatic()) {
   demoDataListContent.listing.items.forEach((item) => {
@@ -65,6 +68,7 @@ let baseData = {
   with_footer: true,
   with_sidebar: true,
   ...demoData,
+  ...dataListingTemplate,
 };
 
 const dataListingContent = {
@@ -77,6 +81,7 @@ const dataListingMember = {
   page_title: "Group members",
   ...baseData,
   ...demoDataListMember,
+  listing: dataProfiles,
 };
 
 delete baseData.banner;
