+++ src/compositions/bcl-news/news.story.js
@@ -2,12 +2,14 @@ import { withDesign } from "storybook-addon-designs";
 import { initBadges, correctPaths } from "@openeuropa/bcl-story-utils";
 import isChromatic from "chromatic/isChromatic";
 import header from "@openeuropa/bcl-data-header/data--simple";
-import dataListing from "@openeuropa/bcl-news/data/data_listing.js";
 import footer from "@openeuropa/bcl-data-footer/data";
-import content from "@openeuropa/bcl-news/data/data_content.js";
-import banner from "@openeuropa/bcl-content-banner/data/data.js";
+import dataListingTemplate from "@openeuropa/bcl-data-templates/listing/data--template";
 import listingPage from "@openeuropa/bcl-base-templates/listing-page.html.twig";
 import news from "@openeuropa/bcl-base-templates/content-type.html.twig";
+import banner from "@openeuropa/bcl-content-banner/data/data.js";
+
+import dataListing from "@openeuropa/bcl-news/data/data_listing.js";
+import content from "@openeuropa/bcl-news/data/data_content.js";
 const feedback = `<div class="bg-lighter py-4 mt-md-4-5 mt-4 text-center">Feedback module here</div>`;
 const share = `<div class="bg-gray-200 py-4 text-center">Share module here</div>`;
 
@@ -41,6 +43,7 @@ const demoData = {
 const demoListing = {
   ...baseData,
   ...dataListing,
+  ...dataListingTemplate,
   with_sidebar: true,
 };
