+++ src/compositions/bcl-news/news.test.js
@@ -1,9 +1,11 @@
 import { renderTwigFileAsNode } from "@openeuropa/bcl-test-utils";
-import dataListing from "@openeuropa/bcl-news/data/data_listing.js";
 import header from "@openeuropa/bcl-data-header/data";
 import footer from "@openeuropa/bcl-data-footer/data";
-import content from "@openeuropa/bcl-news/data/data_content.js";
 import banner from "@openeuropa/bcl-content-banner/data/data.js";
+import dataListingTemplate from "@openeuropa/bcl-data-templates/listing/data--template";
+
+import dataListing from "@openeuropa/bcl-news/data/data_listing.js";
+import content from "@openeuropa/bcl-news/data/data_content.js";
 
 const feedback = `<div class="bg-lighter py-4 mt-md-4-5 mt-4 text-center">Feedback module here</div>`;
 const share = `<div class="bg-gray-200 py-4 text-center">Share module here</div>`;
@@ -29,6 +31,7 @@ const demoData = {
 const demoListing = {
   ...baseData,
   ...dataListing,
+  ...dataListingTemplate,
   with_sidebar: true,
 };
