+++ src/compositions/bcl-person/person.story.js
@@ -1,11 +1,14 @@
 import { withDesign } from "storybook-addon-designs";
 import { initBadges, correctPaths } from "@openeuropa/bcl-story-utils";
 import header from "@openeuropa/bcl-data-header/data--simple";
+import footer from "@openeuropa/bcl-data-footer/data";
+import dataListingTemplate from "@openeuropa/bcl-data-templates/listing/data--template";
+import dataProfiles from "@openeuropa/bcl-data-templates/listing/data--profiles";
+import listingPage from "@openeuropa/bcl-base-templates/listing-page.html.twig";
+
 import dataListing from "@openeuropa/bcl-person/data/data--listing";
 import dataDetails from "@openeuropa/bcl-person/data/data--details";
-import footer from "@openeuropa/bcl-data-footer/data";
 import detailsPage from "@openeuropa/bcl-person/person.html.twig";
-import listingPage from "@openeuropa/bcl-base-templates/listing-page.html.twig";
 
 const baseData = {
   content_type: "listing",
@@ -24,7 +27,9 @@ const demoDetails = {
 
 const demoListing = {
   ...baseData,
+  ...dataListingTemplate,
   ...dataListing,
+  listing: dataProfiles,
 };
 
 export default {
