+++ src/compositions/bcl-search/search.story.js
@@ -5,11 +5,13 @@ import {
 } from "@openeuropa/bcl-story-utils";
 import isChromatic from "chromatic/isChromatic";
 import { withDesign } from "storybook-addon-designs";
-import demoData from "@openeuropa/bcl-search/dataSearch.js";
-import search from "@openeuropa/bcl-search/search.html.twig";
 import header from "@openeuropa/bcl-data-header/data--no-form";
 import footer from "@openeuropa/bcl-data-footer/data";
+import dataListingTemplate from "@openeuropa/bcl-data-templates/listing/data--template";
 import drupalAttribute from "drupal-attribute";
+
+import demoData from "@openeuropa/bcl-search/data/dataSearch.js";
+import search from "@openeuropa/bcl-search/search.html.twig";
 
 if (isChromatic()) {
   demoData.listing.items.forEach((item) => {
@@ -31,6 +33,7 @@ const dataListing = {
   with_header: true,
   with_footer: true,
   with_sidebar: true,
+  ...dataListingTemplate,
   ...demoData,
 };
