+++ src/compositions/bcl-user/data/data-user-view-compact.js
