+++ src/data/footer/data.js
@@ -1,5 +1,3 @@
-const drupalAttribute = require("drupal-attribute");
-
 module.exports = {
   rows: [
     {
@@ -42,7 +40,6 @@ module.exports = {
                     path: "/icons.svg",
                     name: "facebook",
                     size: "xs",
-                    attributes: new drupalAttribute().addClass(["me-2"]),
                   },
                 },
                 {
@@ -52,7 +49,6 @@ module.exports = {
                     path: "/icons.svg",
                     name: "twitter",
                     size: "xs",
-                    attributes: new drupalAttribute().addClass(["me-2"]),
                   },
                 },
                 {
@@ -62,7 +58,6 @@ module.exports = {
                     name: "linkedin",
                     path: "/icons.svg",
                     size: "xs",
-                    attributes: new drupalAttribute().addClass(["me-2"]),
                   },
                 },
               ],
