+++ src/data/templates/listing/data--profiles.js
@@ -1,0 +1,87 @@
+module.exports = {
+  variant: "default-1-col",
+  remove_horizontal_spacer: true,
+  items: [
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content: "<a href='/example.html' class='standalone'>Stefan Mayer</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=1",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content: "<a href='/example.html' class='standalone'>Adjur Tichar</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=2",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content: "<a href='/example.html' class='standalone'>Szép Mara</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=3",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content:
+          "<a href='/example.html' class='standalone'>Daniel Cardoso Rocha</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=4",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content:
+          "<a href='/example.html' class='standalone'>Aurelia Panicucci</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=5",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+    {
+      extra_classes_header: "bg-transparent pt-0 pb-0 border-0",
+      title: {
+        content:
+          "<a href='/example.html' class='standalone'>Felicyta Kowalska</a>",
+      },
+      content:
+        "<span class='text-muted me-4-5'>Organisation</span><span class='text-muted me-2'>Position</span>",
+      image: {
+        path: "https://picsum.photos/160?random=6",
+        alt: "alt img",
+        position: "top",
+      },
+    },
+  ],
+};
