+++ src/data/templates/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@openeuropa/bcl-data-templates",
+  "author": "European Commission",
+  "license": "EUPL-1.2",
+  "version": "0.20.0",
+  "description": "OE templates data",
+  "publishConfig": {
+    "access": "public"
+  },
+  "repository": {
+    "type": "git",
+    "url": ""
+  },
+  "bugs": {
+    "url": ""
+  },
+  "homepage": "",
+  "keywords": [
+    "oe",
+    "component-library",
+    "design-system"
+  ]
+}
